refactor(ui): split _template.py god-file into _styles + _scripts (#236)
The UI viewer's get_html() was a single 1548-line f-string with the
stylesheet and client-side JS inlined, every literal brace doubled
({{ }} / ${{ }}) to survive f-string parsing — the "escape hell" the
QC audit (#236, domain F) flagged.

Extract the CSS and JS verbatim into sibling _styles.py (CSS_BLOCK) and
_scripts.py (JS_BLOCK) modules as plain (non-f) strings, so they carry
real CSS `{ }` and JS `${...}` template literals with no brace doubling.
get_html() now just assembles the two blocks and interpolates the single
(HTML-escaped) title. _template.py: 1548 -> 214 lines.

Purely internal — get_html() output is byte-for-byte identical, verified
across default / XSS-escaping / plain titles. New test_ui_template.py
locks the assembly contract (both blocks embedded verbatim, braces
unescaped, title the only escaped interpolation point).

Full suite: 1297 passed, 2 skipped. mypy strict + ruff clean.
